collect xlsx/xls

+++ Collect data/docx_doc.cs
@@ -8,7 +8,7 @@ using System.Threading.Tasks;
 namespace Collect_data
 {
     class docx_doc
-    {    
+    {
         public static void Docx_doc()
         {
             DirectoryInfo targetFolder = new DirectoryInfo("C:\\tmp");
@@ -34,7 +34,7 @@ namespace Collect_data
                     string resultDocuments = Path.GetFileName(items);
                     string copyFile = Path.Combine("C:\\tmp\\Documents", resultDocuments);
                     File.Copy(items, copyFile, true);
-                }
+                }                         
             }
 
             foreach (var items in file_paths.files_docx_downloads)
@@ -45,8 +45,8 @@ namespace Collect_data
 
                     string resultDownloads = Path.GetFileName(items);
                     string copyFiles = Path.Combine("C:\\tmp\\Downloads", resultDownloads);
-                    File.Copy(items,copyFiles, true);              
-                }   
+                    File.Copy(items, copyFiles, true);
+                }                                  
             }
 
             foreach (var items in file_paths.files_docx_C)
@@ -57,39 +57,56 @@ namespace Collect_data
 
                     string resultC = Path.GetFileName(items);
                     string copyFiles = Path.Combine("C:\\tmp\\disk_C", resultC);
-                    File.Copy(items, copyFiles, true);                
-                }
+                    File.Copy(items, copyFiles, true);
+                }                                                
             }
 
             foreach (var items in file_paths.files_doc_desktop)
             {
                 if (targetFolder.Exists)
                 {
+                    Directory.CreateDirectory("C:\\tmp\\Desktop");
+
                     string resultDesktop = Path.GetFileName(items);
                     string copyFile = Path.Combine("C:\\tmp\\Desktop", resultDesktop);
                     File.Copy(items, copyFile, true);
-                }
+                }                                   
             }
 
             foreach (var items in file_paths.files_doc_documets)
             {
-                string resultDocuments = Path.GetFileName(items);
-                string copyFile = Path.Combine("C:\\tmp\\Documents", resultDocuments);
-                File.Copy(items, copyFile, true);
+                if (targetFolder.Exists)
+                {
+                    Directory.CreateDirectory("C:\\tmp\\Documents");
+
+                    string resultDocuments = Path.GetFileName(items);
+                    string copyFile = Path.Combine("C:\\tmp\\Documents", resultDocuments);
+                    File.Copy(items, copyFile, true);
+                }       
             }
 
             foreach (var items in file_paths.files_doc_downloads)
             {
-                string resultDownloads = Path.GetFileName(items);
-                string copyFiles = Path.Combine("C:\\tmp\\Downloads", resultDownloads);
-                File.Copy(items, copyFiles, true);
+                if (targetFolder.Exists)
+                {
+                    Directory.CreateDirectory("C:\\tmp\\Downloads");
+
+                    string resultDownloads = Path.GetFileName(items);
+                    string copyFiles = Path.Combine("C:\\tmp\\Downloads", resultDownloads);
+                    File.Copy(items, copyFiles, true);
+                }    
             }
 
             foreach (var items in file_paths.files_doc_C)
             {
-                string resultC = Path.GetFileName(items);
-                string copyFiles = Path.Combine("C:\\tmp\\disk_C", resultC);
-                File.Copy(items, copyFiles, true);
+                if (targetFolder.Exists)
+                {
+                    Directory.CreateDirectory("C:\\tmp\\disk_C");
+
+                    string resultC = Path.GetFileName(items);
+                    string copyFiles = Path.Combine("C:\\tmp\\disk_C", resultC);
+                    File.Copy(items, copyFiles, true);
+                } 
             }
         }
     }
